Changes proposed from https://github.com/OLEGSHA/PiwcsPrwModel/pull/1#discussion_r1577679369
- Rename `isLocCorrect(const Model& model, IdRef id)` to `isLocallyCorrect(const Model& model, IdRef id`
- Move `isLocallyCorrect` and `isCorrect` to correctness.cpp
- Introduce private header for purpose of linking test executable against `isLocallyCorrect`
